requirements.txt
aes==1.2.0
asgiref==3.5.2
cachetools==5.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
checksum==1.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-adminlte-3==0.1.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-jazzmin==2.6.0
django-localflavor==3.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
numpy==1.24.0
oauthlib==3.2.2
pandas==1.5.2
Pillow==9.3.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.16.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-stdnum==1.18
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
uuid==1.30
